Add routes for user-meeting associations

+++ server/app/controllers/meetings_controller.rb
@@ -5,29 +5,50 @@ class MeetingsController < ApplicationController
     if !(logged_in? && current_user)
       render json: {
         status: 500,
-        errors: ["user not logged in"],
+        errors: ["User not logged in"],
       }
     end
   end
 
+  def confirm_requester_is_owner_or_admin(owner_id)
+    if !(current_user.id == owner_id or current_user.usertype == "admin")
+      render json: {
+        status: 500,
+        errors: ["User does not have previleges for requested action"],
+      }
+      return false
+    end
+    return true
+  end
+
+  def confirm_requester_is_admin
+    if current_user.usertype != "admin"
+      render json: {
+        status: 500,
+        errors: ["User does not have previleges for requested action"],
+      }
+      return false
+    end
+    return true
+  end
+
   # GET /meetings
   def index
-    @meetings = Meeting.all
-    if @meetings
-      render json: {
-               meetings: @meetings,
-             }
-    else
-      render json: {
-               status: 500,
-               errors: ["no meetings found"],
-             }
+    if !confirm_requester_is_admin
+      return
     end
+
+    render json: {
+             meetings: Meeting.all,
+           }
   end
 
   # GET /meetings/1
   def show
     @meeting = Meeting.find_by_id(params[:id])
+    if @meeting and !confirm_requester_is_owner_or_admin(@meeting.owner.id)
+      return
+    end
     if @meeting
       render json: {
                meeting: @meeting,
@@ -35,13 +56,17 @@ class MeetingsController < ApplicationController
     else
       render json: {
                status: 500,
-               errors: ["meeting not found"],
+               errors: ["Meeting not found"],
              }
     end
   end
 
   # POST /meetings/
   def create
+    # TODO: handle creation by company representatives
+    if !confirm_requester_is_admin
+      return
+    end
     params = meeting_params.to_h
     # If owner key is not provided, use the owner key provided by the requester
     if !params.key?("owner_id")
@@ -65,6 +90,9 @@ class MeetingsController < ApplicationController
   # PUT /meetings/1/
   def update
     @meeting = Meeting.find(params[:id])
+    if @meeting and !confirm_requester_is_owner_or_admin(@meeting.owner_id)
+      return
+    end
     if @meeting.update(meeting_params)
       render json: {
                status: :created,
@@ -81,9 +109,12 @@ class MeetingsController < ApplicationController
   # DELETE /meetings/1/
   def destroy
     @meeting = Meeting.find(params[:id])
+    if @meeting and !confirm_requester_is_owner_or_admin(@meeting.owner_id)
+      return
+    end
     if @meeting.destroy
       render json: {
-               status: :destroyed,
+               status: 200,
                meeting: @meeting,
              }
     else
